fix util test
