+++ src/utils.ts
@@ -1,6 +1,6 @@
 import {whiteBright} from 'cli-color'
 import {deburr, isPlainObject, mapValues, trim, upperFirst} from 'lodash'
-import {basename, extname} from 'path'
+import {basename, extname, join} from 'path'
 import {JSONSchema} from './types/JSONSchema'
 
 // TODO: pull out into a separate package
@@ -220,4 +220,27 @@ export function escapeBlockComment(schema: JSONSchema) {
       schema[key] = schema[key]!.replace(/\*\//g, replacer)
     }
   }
+}
+
+/*
+the following logic determines the out path by comparing the in path to the users specified out path.
+For example, if input directory MultiSchema looks like:
+  MultiSchema/foo/a.json
+  MultiSchema/bar/fuzz/c.json
+  MultiSchema/bar/d.json
+And the user wants the outputs to be in MultiSchema/Out, then this code will be able to map the inner directories foo, bar, and fuzz into the intended Out directory like so:
+  MultiSchema/Out/foo/a.json
+  MultiSchema/Out/bar/fuzz/c.json
+  MultiSchema/Out/bar/d.json
+*/
+export function pathTransform(o: string, i: string): string {
+  const outPathList = o.split('/')
+  const inPathList = i.split('/')
+
+  const intersection = outPathList.filter(x => inPathList.includes(x))
+  const symmetricDifference = outPathList
+    .filter(x => !inPathList.includes(x))
+    .concat(inPathList.filter(x => !outPathList.includes(x)))
+
+  return join(...intersection, ...symmetricDifference)
 }
