+++ test/test.ts
@@ -2,6 +2,7 @@ import {run as runCLITests} from './testCLI'
 import {run as runCompileFromFileTests} from './testCompileFromFile'
 import {hasOnly, run as runE2ETests} from './testE2E'
 import {run as runNormalizerTests} from './testNormalizer'
+import {run as runUtilsTests} from './testUtils'
 
 runE2ETests()
 
@@ -9,4 +10,5 @@ if (!hasOnly()) {
   runCompileFromFileTests()
   runCLITests()
   runNormalizerTests()
+  runUtilsTests()
 }
